fix: improvements from agent branch
